MVP2 피드백, 스프린트3 정보수정제안

+++ app/src/main/java/com/jeju/nanaland/util/daterangecalendar/customviews/CustomDateView.kt
@@ -74,7 +74,7 @@ class CustomDateView @JvmOverloads constructor(
     override var stripColor: Int = defCalendarStyleAttr.rangeStripColor
 
     private val mViewClickListener = OnClickListener {
-        val key = it.tag as Int
+        val key = it.tag as Long
         if (onDateClickListener != null) {
             val selectedCal = Calendar.getInstance()
             var date = Date()
